+++ temba/flows/views.py
@@ -35,8 +35,8 @@ from temba.flows.models import Flow, FlowRevision, FlowRun, FlowSession, FlowSta
 from temba.flows.tasks import update_session_wait_expires
 from temba.ivr.models import Call
 from temba.orgs.models import IntegrationType, Org
-from temba.orgs.views import BaseExportView, DependencyDeleteModal
-from temba.orgs.views.base import BaseListView, BaseMenuView
+from temba.orgs.views import DependencyDeleteModal
+from temba.orgs.views.base import BaseExportModal, BaseListView, BaseMenuView
 from temba.orgs.views.mixins import BulkActionMixin, OrgObjPermsMixin, OrgPermsMixin
 from temba.triggers.models import Trigger
 from temba.utils import analytics, gettext, json, languages, on_transaction_commit
@@ -1153,8 +1153,8 @@ class FlowCRUDL(SmartCRUDL):
         def derive_initial(self):
             return {"language": self.po_info.language_code if self.po_info else ""}
 
-    class ExportResults(BaseExportView):
-        class Form(BaseExportView.Form):
+    class ExportResults(BaseExportModal):
+        class Form(BaseExportModal.Form):
             flows = forms.ModelMultipleChoiceField(
                 Flow.objects.none(), required=True, widget=forms.MultipleHiddenInput()
             )
